Refactor: Move from_pretrained to train_utils and fix circular dependencies
