+++ MaxText/__init__.py
@@ -29,32 +29,6 @@ from MaxText import train_utils
 from MaxText import pyconfig
 from MaxText.layers import models
 from MaxText import dpo_utils
-import jax
-from jax.sharding import Mesh
+from MaxText.train_utils import from_pretrained
 
 Transformer = models.Transformer
-
-
-def from_pretrained(
-    config: pyconfig.HyperParameters,
-    devices: Sequence[jax.Device] | None = None,
-) -> Transformer:
-  """Load a pretrained MaxText model from checkpoint.
-
-  This function loads a model from a checkpoint.
-
-  Args:
-      config: Config object.
-
-  Returns:
-      Transformer: The loaded model instance (only the model)
-
-  Example:
-      model = from_pretrained(config)
-  """
-  devices_array = maxtext_utils.create_device_mesh(config, devices)
-  mesh = Mesh(devices_array, config.mesh_axes)
-  model = train_utils.create_model(config, mesh)
-
-  # Return only the model
-  return model
